refactor: rename Go binary orchard → orchard-daemon (ADR-013 step 5)
Renames cmd/orchard/ → cmd/orchard-daemon/ and updates the Makefile,
README, and Go E2E tests. The dispatcher (PR #444) routes
`orchard daemon ...` to this binary; users no longer invoke it
directly.

Install layout (post-merge):
  /usr/local/bin/orchard           → dispatcher (only binary users invoke)
  /usr/local/bin/orchard-tui       → TUI helper
  /usr/local/bin/orchard-daemon    → Go daemon (was orchard)
  /usr/local/bin/orchard-worktree  → worktree mutation CLI

Makefile changes:
- daemon target now builds bin/orchard-daemon (was bin/orchard).
- New install-dispatcher target: target/release/orchard → /usr/local/bin/orchard.
- New install-worktree-cli target: target/release/orchard-worktree → /usr/local/bin/orchard-worktree.
- `make install` runs all four install-* targets.
- Top-of-Makefile docstring updated to reflect the four-binary layout.

cobra Use field changed from "orchard" to "orchard-daemon" so --help
output reflects the dispatched form. Stderr prefix similarly updated.

Tests:
- go test ./... PASSES (incl. E2E tests that build the binary from the
  renamed path).
- cargo test --workspace: only #439 pre-existing flakes failed; new code
  unaffected.
- ./bin/orchard-daemon --version and --help work as expected.

After this lands, step 6 (namespace remaining flat verbs) is the last
ADR-013 sequence item.
